WIP: Added Action Sheet and AnalyticalCard Test
