+++ packages/main/src/components/AnalyticalCard/AnalyticalCard.test.tsx
@@ -1,14 +1,10 @@
 import { mountThemedComponent } from '@shared/tests/utils';
-import { expect, use } from 'chai';
-import { matchSnapshot } from 'chai-karma-snapshot';
 import React from 'react';
 import { AnalyticalCard } from '../../lib/AnalyticalCard';
 import { AnalyticalCardHeader } from '../../lib/AnalyticalCardHeader';
 import { DeviationIndicator } from '../../lib/DeviationIndicator';
 import { Text } from '../../lib/Text';
 import { ValueState } from '../../lib/ValueState';
-
-use(matchSnapshot);
 
 const Header = (
   <AnalyticalCardHeader
@@ -28,12 +24,12 @@ const Header = (
 
 // TODO Add more tests
 describe('Analytical Card', () => {
-  it('Render without Crashing', () => {
+  test('Render without Crashing', () => {
     const wrapper = mountThemedComponent(
       <AnalyticalCard header={Header}>
         <Text>I'm a content!</Text>
       </AnalyticalCard>
     );
-    expect(wrapper.debug()).to.matchSnapshot();
+    expect(wrapper.render()).toMatchSnapshot();
   });
 });
